move RegisterForm.tsx to auth folder
