+++ src/app/_components/auth/LoginForm.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from "../common/RegisterForm";
+import RegisterForm from "./RegisterForm";
 import { loginFormData } from "@/app/_lib/_data/RegisterFormData";
 import { loginFormInputs } from "@/app/_lib/_data/RegisterFormData";
 import registerServerAction from "@/app/_actions/registerAction";
